Size..

+++ 0x00-hello_world/6-size.c
@@ -1,21 +1,21 @@
-#include <stdio.d>
+#include <stdio.h>
 /**
- * main - Write a program that prints the size of various types
- *
- * Return: 0
- */
+* main - Write a program that prints the size of various types
+*
+* Return: 0
+*/
 int main(void)
 {
 	char a;
-        int b;
-        long int c;
-        long long int d;
-        float e;
-	
-	printf("Size of a char: %lu byte(s)\n", (unsigned int)sizeof(a));
-	printf("Size of an int: %lu byte(s)\n", (unsigned int)sizeof(b));
-	printf("Size of a long int: %lu byte(s)\n", (unsigned int)sizeof(c));
-	printf("Size of a long long int: %lu byte(s)\n", (unsigned int)sizeof(d));
-	printf("Size of a float: %lu byte(s)\n", (unsigned int)sizeof(e));
+	int b;
+	long int c;
+	long long int d;
+	float e;
+
+	printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(a));
+	printf("Size of an int: %lu byte(s)\n", (unsigned long)sizeof(b));
+	printf("Size of a long int: %lu byte(s)\n", (unsigned long)sizeof(c));
+	printf("Size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(d));
+	printf("Size of a float: %lu byte(s)\n", (unsigned long)sizeof(e));
 	return (0);
 }
